Ports Skrell to JollyStation (#28)
* Trying to port Skrell

The start of trying to port Skrell, so far the race shows up, added them to the roundstart config, but their hair does not show up. Cannot figure out what is up right now.

* Make Skrell Light Drinkers

Gives them the trait inherently, I figure this is the best way to handle it to avoid just having two checks for a similar thing.

* Finishes up porting the Skrell

Fixed up a few issues with them, added blood type, blood bag, and got the language working with a custom tongue for Skrell.

* Move Copper on_life to Skrell.dm

See title

* Fix Merge Conflict

Fixes the dme?